chore: Added desafio6.sql command for resolution of requirement number 6
fix: changed subscription_value colum type to DECIMAL on subscriptions table

+++ desafio1.sql
@@ -4,7 +4,7 @@ CREATE DATABASE SpotifyClone ;
 CREATE TABLE SpotifyClone.subscriptions (
   subscription_id INT NOT NULL AUTO_INCREMENT,
   subscription_type VARCHAR(45) NULL,
-  subscription_value FLOAT(3,2) NULL,
+  subscription_value DECIMAL(3,2) NULL,
   PRIMARY KEY (subscription_id))
 ENGINE = InnoDB;
